optimize ConfigModifyVerifyService
