+++ hippo4j-config/src/main/java/cn/hippo4j/config/controller/ConfigVerifyController.java
@@ -43,4 +43,5 @@ public class ConfigVerifyController {
                         () -> modifyVerifyService.rejectModification(reqDTO.getId(), reqDTO.getThreadPoolParameterInfo()));
         return Results.success();
     }
+
 }
