+++ hippo4j-config/src/main/java/cn/hippo4j/config/service/biz/impl/ThreadPoolInstanceConfigModifyVerifyServiceImpl.java
@@ -36,8 +36,8 @@ public class ThreadPoolInstanceConfigModifyVerifyServiceImpl extends AbstractCon
     }
 
     @Override
-    @Transactional(rollbackFor = Exception.class)
-    public void acceptModification(Long id, ThreadPoolParameterInfo poolParameterInfo) {
+    protected void updateThreadPoolParameter(ThreadPoolParameterInfo poolParameterInfo) {
+
     }
 
 }
