+++ hippo4j-config/src/main/java/cn/hippo4j/config/service/biz/impl/ThreadPoolManageConfigModifyVerifyServiceImpl.java
@@ -18,21 +18,14 @@
 package cn.hippo4j.config.service.biz.impl;
 
 import cn.hippo4j.common.constant.ConfigModifyTypeConstants;
-import cn.hippo4j.common.enums.VerifyEnum;
 import cn.hippo4j.common.model.ThreadPoolParameterInfo;
-import cn.hippo4j.common.toolkit.UserContext;
-import cn.hippo4j.config.mapper.HisConfigVerifyMapper;
-import cn.hippo4j.config.model.HisConfigVerifyInfo;
 import cn.hippo4j.config.model.biz.threadpool.ThreadPoolSaveOrUpdateReqDTO;
 import cn.hippo4j.config.service.biz.ThreadPoolService;
 import cn.hippo4j.config.toolkit.BeanUtil;
-import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
-import java.util.Date;
 
 @Slf4j
 @Service
@@ -47,16 +40,7 @@ public class ThreadPoolManageConfigModifyVerifyServiceImpl extends AbstractConfi
     }
 
     @Override
-    @Transactional(rollbackFor = Exception.class)
-    public void acceptModification(Long id, ThreadPoolParameterInfo poolParameterInfo) {
-        LambdaUpdateWrapper<HisConfigVerifyInfo> updateWrapper = new LambdaUpdateWrapper<HisConfigVerifyInfo>()
-                .eq(HisConfigVerifyInfo::getId, id)
-                .set(HisConfigVerifyInfo::getVerifyStatus, VerifyEnum.VERIFY_ACCEPT.getVerifyStatus())
-                .set(HisConfigVerifyInfo::getGmtVerify, new Date())
-                .set(HisConfigVerifyInfo::getVerifyUser, UserContext.getUserName());
-
-        hisConfigVerifyMapper.update(null, updateWrapper);
-
+    public void updateThreadPoolParameter(ThreadPoolParameterInfo poolParameterInfo) {
         ThreadPoolSaveOrUpdateReqDTO reqDTO = BeanUtil.convert(poolParameterInfo, ThreadPoolSaveOrUpdateReqDTO.class);
         threadPoolService.saveOrUpdateThreadPoolConfig(null, reqDTO);
     }
