+++ hippo4j-config/src/main/java/cn/hippo4j/config/service/biz/impl/WebThreadPoolConfigModifyVerifyServiceImpl.java
@@ -34,7 +34,7 @@ public class WebThreadPoolConfigModifyVerifyServiceImpl extends AbstractConfigMo
     }
 
     @Override
-    public void acceptModification(Long id, ThreadPoolParameterInfo poolParameterInfo) {
-    }
+    protected void updateThreadPoolParameter(ThreadPoolParameterInfo poolParameterInfo) {
 
+    }
 }
